Remove webpack-dev-server
Severe security problems, and it is not needed since it only is used locally.

Here, it is replaced by a watch script, and relative paths to assets and resources so that index.html can be viewed directly in the browser instead of via a dev server.

Also, some minor production preparing stuffs (hiding elements not yet ready) and some ugly white space inconsistecy fixes.